requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.3
Brlapi==0.8.3
cachetools==5.3.1
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
colorgram.py==1.2.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.7
django-extensions==3.2.3
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
duplicity==0.8.21
et-xmlfile==1.1.0
Faker==20.1.0
fasteners==0.14.1
Flask==3.0.0
flatbuffers==23.5.26
frozendict==2.3.8
frozenlist==1.4.1
future==0.18.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
html5lib==1.1
httplib2==0.20.2
idna==3.6
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
instagrapi==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
multidict==6.0.4
multitasking==0.0.11
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paramiko==2.9.3
peewee==3.17.0
pexpect==4.8.0
pillow==10.2.0
pipreqs==0.4.13
pluggy==1.3.0
proglog==0.1.10
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.18.0
pycups==2.0.1
pydantic==1.10.9
pygame==2.5.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytest==7.4.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.1

termcolor==2.3.0
threadpoolctl==3.2.0
tk==0.1.0
tqdm==4.66.1
ttkthemes==3.2.2
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.1.0
usb-creator==0.3.7
validator==0.7.1
validators==0.22.0
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.9.4
yfinance==0.2.31
zipp==1.0.0
